feat : 온라인 심리검사결과 리스트 출력 컨테이너 구성함
